Imported vanilla DSpace 1.6.2.

+++ dspace-api/src/main/java/org/dspace/checker/BitstreamInfoDAO.java
@@ -1,6 +1,5 @@
 /*
- * Copyright (c) 2004-2005, Hewlett-Packard Company and Massachusetts
- * Institute of Technology.  All rights reserved.
+ * Copyright (c) 2002-2009, The DSpace Foundation.  All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are
@@ -13,8 +12,7 @@
  * notice, this list of conditions and the following disclaimer in the
  * documentation and/or other materials provided with the distribution.
  *
- * - Neither the name of the Hewlett-Packard Company nor the name of the
- * Massachusetts Institute of Technology nor the names of their
+ * - Neither the name of the DSpace Foundation nor the names of its
  * contributors may be used to endorse or promote products derived from
  * this software without specific prior written permission.
  *
@@ -274,7 +272,7 @@ public final class BitstreamInfoDAO extends DAOSupport
             if (rs.next())
             {
                 info = new BitstreamInfo(rs.getBoolean("deleted"), rs
-                        .getInt("store_number"), rs.getInt("size_bytes"), rs
+                        .getInt("store_number"), rs.getLong("size_bytes"), rs
                         .getString("short_description"), rs
                         .getInt("bitstream_id"), rs
                         .getString("user_format_description"), rs
@@ -546,7 +544,7 @@ public final class BitstreamInfoDAO extends DAOSupport
     /**
      * Get the bitstream ids for a given collection
      * 
-     * @param itemId
+     * @param collectionId The id of the collection
      * @return the list of bitstream ids for this item
      */
     public List getCollectionBitstreams(int collectionId)
@@ -589,7 +587,7 @@ public final class BitstreamInfoDAO extends DAOSupport
     /**
      * Get the bitstream ids for a given community
      * 
-     * @param itemId
+     * @param communityId the community id
      * @return the list of bitstream ids for this item
      */
     public List getCommunityBitstreams(int communityId)
